requirements.txt
# ============================================================
# Research Platform - Backend Requirements
# ============================================================
# Install with:
#   python -m venv .venv
#   .venv\Scripts\activate
#   pip install -r requirements.txt
# ============================================================

# --- Web Framework & Server ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
starlette==0.35.1

# --- Data Validation ---
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic-core==2.14.6

# --- Machine Learning & Embeddings ---
torch==2.6.0
torchvision==0.21.0
transformers==4.36.2
sentence-transformers==2.3.1
safetensors==0.7.0
tokenizers==0.15.2
huggingface_hub==0.36.2
sentencepiece==0.2.1

# --- Vector Search ---
faiss-cpu==1.13.2

# --- Scientific Computing ---
numpy==1.26.3
scipy==1.15.3
scikit-learn==1.4.0
scikit-image==0.25.2

# --- Document Parsing & AI ---

pymupdf==1.24.11
pypdf==4.3.1
pypdfium2==4.30.0
python-docx==1.2.0
python-pptx==1.0.2
easyocr==1.7.2

# --- LLM API ---
groq==0.9.0

# --- Authentication ---
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0

# --- HTTP & Networking ---
requests==2.31.0
httpx==0.27.0
httpcore==1.0.9
anyio==4.12.1

# --- Data Processing ---
pandas==2.1.4
pyarrow==16.1.0
feedparser==6.0.10
xmltodict==0.13.0
beautifulsoup4==4.14.3
lxml==6.0.2

# --- Image Processing ---
pillow==10.4.0
opencv-python-headless==4.11.0.86

# --- Utilities ---
python-dotenv==1.0.0
tqdm==4.66.1
python-dateutil==2.8.2
ratelimit==2.2.1
tabulate==0.9.0
xlsxwriter==3.2.9
rich==13.9.4
PyYAML==6.0.3
sympy==1.13.1
regex==2026.2.28
filelock==3.25.0
packaging==26.0
typing_extensions==4.15.0

# --- Windows Specific (remove if running on Mac/Linux) ---
pywin32==307
colorama==0.4.6